+++ tiger-common/src/main/java/de/gematik/rbellogger/util/CryptoLoader.java
@@ -4,6 +4,8 @@
 
 package de.gematik.rbellogger.util;
 
+import de.gematik.test.tiger.common.exceptions.TigerPkiException;
+import de.gematik.test.tiger.common.pki.TigerPkiIdentity;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import org.bouncycastle.jce.provider.BouncyCastleProvider;
@@ -38,9 +40,9 @@ public class CryptoLoader {
                 return (X509Certificate) p12.getCertificate(alias);
             }
         } catch (final IOException | KeyStoreException | NoSuchAlgorithmException | CertificateException e) {
-            throw new RuntimeException(e);
+            throw new TigerPkiException("Exception while trying to extract certificate from p12!", e);
         }
-        throw new RuntimeException("Could not find certificate in P12-File");
+        throw new TigerPkiException("Could not find certificate in P12-File");
     }
 
     public static X509Certificate getCertificateFromPem(final byte[] crt) {
@@ -49,15 +51,15 @@ public class CryptoLoader {
             final InputStream in = new ByteArrayInputStream(crt);
             final X509Certificate x509Certificate = (X509Certificate) certFactory.generateCertificate(in);
             if (x509Certificate == null) {
-                throw new RuntimeException("Error while loading certificate!");
+                throw new TigerPkiException("Error while loading certificate (null)!");
             }
             return x509Certificate;
         } catch (final CertificateException ex) {
-            throw new RuntimeException("Error while loading certificate!", ex);
+            throw new TigerPkiException("Error while loading certificate!", ex);
         }
     }
 
-    public static RbelPkiIdentity getIdentityFromP12(final byte[] p12FileContent, final String p12Password) {
+    public static TigerPkiIdentity getIdentityFromP12(final byte[] p12FileContent, final String p12Password) {
         try {
             final KeyStore p12 = KeyStore.getInstance("pkcs12", BOUNCY_CASTLE_PROVIDER);
             p12.load(new ByteArrayInputStream(p12FileContent), p12Password.toCharArray());
@@ -69,16 +71,16 @@ public class CryptoLoader {
                 if (privateKey == null) {
                     continue;
                 }
-                return new RbelPkiIdentity(certificate, privateKey, Optional.of(alias));
+                return new TigerPkiIdentity(certificate, privateKey, Optional.of(alias));
             }
         } catch (final IOException | KeyStoreException | NoSuchAlgorithmException
             | UnrecoverableKeyException | CertificateException e) {
-            throw new RuntimeException(e);
+            throw new TigerPkiException("Exception while trying to extract identity from p12!", e);
         }
-        throw new RuntimeException("Could not find certificate in P12-File");
+        throw new TigerPkiException("Could not find certificate in P12-File");
     }
 
-    public static RbelPkiIdentity getIdentityFromPemAndPkcs8(final byte[] certificateData, final byte[] keyBytes) {
+    public static TigerPkiIdentity getIdentityFromPemAndPkcs8(final byte[] certificateData, final byte[] keyBytes) {
         try (final ByteArrayInputStream in = new ByteArrayInputStream(keyBytes);
              final InputStreamReader inputStreamReader = new InputStreamReader(in);
              final PemReader pemReader = new PemReader(inputStreamReader);) {
@@ -87,16 +89,16 @@ public class CryptoLoader {
             PemObject pemObject = pemReader.readPemObject();
             byte[] content = pemObject.getContent();
             PKCS8EncodedKeySpec privKeySpec = new PKCS8EncodedKeySpec(content);
-            return RbelPkiIdentity.builder()
+            return TigerPkiIdentity.builder()
                 .certificate(certificate)
                 .privateKey(factory.generatePrivate(privKeySpec))
                 .build();
         } catch (final NoSuchAlgorithmException | InvalidKeySpecException | IOException e) {
-            throw new RuntimeException(e);
+            throw new TigerPkiException("Exception while trying to extract identity from pkcs8!", e);
         }
     }
 
-    public static RbelPkiIdentity getIdentityFromPemAndPkcs1(final byte[] certificateData, final byte[] keyBytes) {
+    public static TigerPkiIdentity getIdentityFromPemAndPkcs1(final byte[] certificateData, final byte[] keyBytes) {
         try (final ByteArrayInputStream in = new ByteArrayInputStream(keyBytes);
              final InputStreamReader inputStreamReader = new InputStreamReader(in);
              final PEMParser pemParser = new PEMParser(inputStreamReader)) {
@@ -105,12 +107,12 @@ public class CryptoLoader {
             KeyPair keyPair = converter.getKeyPair((PEMKeyPair) object);
 
             X509Certificate certificate = getCertificateFromPem(certificateData);
-            return RbelPkiIdentity.builder()
+            return TigerPkiIdentity.builder()
                 .certificate(certificate)
                 .privateKey(keyPair.getPrivate())
                 .build();
         } catch (final IOException e) {
-            throw new RuntimeException(e);
+            throw new TigerPkiException("Exception while trying to extract identity from pkcs1!", e);
         }
     }
 }