+++ tiger-common/src/main/java/de/gematik/test/tiger/common/config/RbelModificationDescription.java
@@ -2,7 +2,7 @@
  * ${GEMATIK_COPYRIGHT_STATEMENT}
  */
 
-package de.gematik.rbellogger.modifier;
+package de.gematik.test.tiger.common.config;
 
 import lombok.*;