+++ tiger-common/src/main/java/de/gematik/test/tiger/common/data/config/tigerProxy/TigerFileSaveInfo.java
@@ -5,11 +5,7 @@
 package de.gematik.test.tiger.common.data.config.tigerProxy;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
-import de.gematik.rbellogger.configuration.RbelFileSaveInfo;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
+import lombok.*;
 
 @Data
 @AllArgsConstructor(onConstructor_ = @JsonIgnore)
@@ -17,16 +13,12 @@ import lombok.NoArgsConstructor;
 @Builder
 public class TigerFileSaveInfo {
 
-    private boolean writeToFile;
     private String sourceFile;
-    private String filename;
-    private boolean clearFileOnBoot;
 
-    public RbelFileSaveInfo toRbelFileSaveInfo() {
-        return RbelFileSaveInfo.builder()
-            .writeToFile(writeToFile)
-            .filename(filename)
-            .clearFileOnBoot(clearFileOnBoot)
-            .build();
-    }
+    @Builder.Default
+    private boolean writeToFile = false;
+    @Builder.Default
+    private String filename = "tiger-proxy.tgr";
+    @Builder.Default
+    private boolean clearFileOnBoot = false;
 }