+++ tiger-proxy/src/main/java/de/gematik/test/tiger/proxy/ITigerProxy.java
@@ -6,7 +6,7 @@ package de.gematik.test.tiger.proxy;
 
 
 import de.gematik.rbellogger.data.RbelElement;
-import de.gematik.rbellogger.modifier.RbelModificationDescription;
+import de.gematik.test.tiger.common.config.RbelModificationDescription;
 import de.gematik.test.tiger.common.data.config.tigerProxy.TigerRoute;
 import java.security.Key;
 import java.util.Deque;