+++ tiger-rbel/src/test/java/de/gematik/rbellogger/util/CryptoLoaderTest.java
@@ -4,6 +4,7 @@
 
 package de.gematik.rbellogger.util;
 
+import de.gematik.test.tiger.common.pki.TigerPkiIdentity;
 import org.apache.commons.io.FileUtils;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -18,7 +19,7 @@ public class CryptoLoaderTest {
     @Test
     @DisplayName("Load ECC Identity from PEM and PKCS8 files")
     public void loadEccIdentityFromp8AndCert() throws IOException {
-        final RbelPkiIdentity pkiIdentity = CryptoLoader.getIdentityFromPemAndPkcs8(
+        final TigerPkiIdentity pkiIdentity = CryptoLoader.getIdentityFromPemAndPkcs8(
             FileUtils.readFileToByteArray(new File("src/test/resources/ecc/cert.pem")),
             FileUtils.readFileToByteArray(new File("src/test/resources/ecc/key.p8"))
         );
@@ -34,7 +35,7 @@ public class CryptoLoaderTest {
     @Test
     @DisplayName("Load RSA Identity from PEM and PKCS8 files")
     public void loadRsaIdentityFromP8AndCert() throws IOException {
-        final RbelPkiIdentity pkiIdentity = CryptoLoader.getIdentityFromPemAndPkcs8(
+        final TigerPkiIdentity pkiIdentity = CryptoLoader.getIdentityFromPemAndPkcs8(
             FileUtils.readFileToByteArray(new File("src/test/resources/rsa/cert.pem")),
             FileUtils.readFileToByteArray(new File("src/test/resources/rsa/key.p8"))
         );
@@ -50,7 +51,7 @@ public class CryptoLoaderTest {
     @Test
     @DisplayName("Load RSA Identity from PEM and PKCS1 files")
     public void loadRsaIdentityFromP1AndCert() throws IOException {
-        final RbelPkiIdentity pkiIdentity = CryptoLoader.getIdentityFromPemAndPkcs1(
+        final TigerPkiIdentity pkiIdentity = CryptoLoader.getIdentityFromPemAndPkcs1(
             FileUtils.readFileToByteArray(new File("src/test/resources/rsa_pkcs1/cert.pem")),
             FileUtils.readFileToByteArray(new File("src/test/resources/rsa_pkcs1/key.pem"))
         );
@@ -66,7 +67,7 @@ public class CryptoLoaderTest {
     @Test
     @DisplayName("Load ECC Identity from P12 file")
     public void loadEccIdentityFromP12() throws IOException {
-        final RbelPkiIdentity pkiIdentity = CryptoLoader.getIdentityFromP12(
+        final TigerPkiIdentity pkiIdentity = CryptoLoader.getIdentityFromP12(
             FileUtils.readFileToByteArray(new File("src/test/resources/idpEnc.p12")),
             "00"
         );
@@ -82,7 +83,7 @@ public class CryptoLoaderTest {
     @Test
     @DisplayName("Load RSA Identity from P12 file")
     public void loadRsaIdentityFromP12() throws IOException {
-        final RbelPkiIdentity pkiIdentity = CryptoLoader.getIdentityFromP12(
+        final TigerPkiIdentity pkiIdentity = CryptoLoader.getIdentityFromP12(
             FileUtils.readFileToByteArray(new File("src/test/resources/rsa.p12")),
             "00"
         );