+++ tiger-test-lib/src/main/java/de/gematik/test/tiger/lib/TigerDirector.java
@@ -157,9 +157,9 @@ public class TigerDirector {
             RbelOptions.deactivateRbelPathDebugging();
         }
         if (libConfig.isRbelAnsiColors()) {
-            RbelOptions.activateAnsiColors();
+            RbelAnsiColors.activateAnsiColors();
         } else {
-            RbelOptions.deactivateAnsiColors();
+            RbelAnsiColors.deactivateAnsiColors();
         }
     }