+++ ReleaseNotes.md
@@ -11,6 +11,8 @@
 * Spring Boot 2.7.5
 
 ## Features
+
+* TGR-741: tiger-commons now has minimal dependencies (no more mock server et al.)
 * TGR-412: ObjectMapper of TigerGlobalConfiguration can now be accessed (and customized)
 
 -------